I realized that the overpotential adjoint wasn't correct, so g_kinetic adjoint is tabled for now until I fix that...but then I noticed overpotential wasn't even working quite right, so I'm trying to fix that and I'm stuck atm. I've tested all three branches and they seem to be working fine in the REPL but the AD version just doesn't converge, even though the non-AD one works fine.
I also decided (at least for now) to specialize `overpotential` to only work on scalar k's since the vector model case is more important for fast phase diagrams. To get vector k's to work, we'd need to allow matrix values for the `guess`, which means `rate_constant` needs to take in matrix voltage values. Not impossible, but not a priority atm.

Signed-off-by: Rachel Kurchin <rkurchin@cmu.edu>